feat(ollama): add embedding node and update embedding node
# Add Ollama Embedding node and update Ollama Embedding node

This commit introduces a new Ollama Embedding node and updates the existing
Ollama Embedding node. The changes are as follows:

1. Added a new `OllamaEmbeddingNode` in `OllamaEmbeddingNode.ts` that provides
   functionality to generate text embeddings using the Ollama API.
2. Updated the existing `ollamaEmbed` node in `index.ts` to use the new
   `OllamaEmbeddingNode` implementation.
3. The new Ollama Embedding node supports both model and text input, allowing
   users to customize the embedding generation process.
4. The node also includes editor definitions for configuring the model and text
   inputs.
5. The node's body display has been updated to show the selected model and text
   (either from input or the default values).

These changes improve the functionality and flexibility of the Ollama
integration within the Rivet platform, allowing users to generate text
embeddings more easily and with more control over the underlying model.
